temp

+++ FileConverter/Structures/StaticData.cs
@@ -5,7 +5,7 @@ namespace FileConverter.Structures
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
     internal struct StaticData
     {
-        public int Id;
+        public ushort Id;
         public StaticTextureInfo ECTexture;
         public StaticTextureInfo CCTexture;
         public RadarColor RadarColor;
